requirements.txt
protobuf==6.31.1
pytubefix==9.2.0
google-generativeai
pandas
openai-whisper